Extract row mappers to their own package
