+++ src/functions/pollJournals/framework/repo/mysql/advance-test-slots-repository.ts
@@ -2,6 +2,7 @@ import * as mysql from 'mysql';
 import { query } from '../database';
 import { ExaminerAdvanceTestSlot } from '../../../domain/examiner-advance-test-slot';
 import * as moment from 'moment';
+import { mapRow } from './row-mappers/advance-test-slot-row-mapper';
 
 export const getAdvanceTestSlots = async (connectionPool: mysql.Pool):
   Promise<ExaminerAdvanceTestSlot[]> => {
@@ -25,34 +26,4 @@ export const getAdvanceTestSlots = async (connectionPool: mysql.Pool):
     /* tslint:enable */
   );
   return res.map(mapRow);
-};
-
-interface AdvanceTestSlotRow {
-  individual_id: number;
-  slot_id: number;
-  start_time: string;
-  minutes: number;
-  tc_id: number;
-  tc_name: string;
-  tc_cost_centre_code: string;
-  short_vst_desc: string;
-}
-
-const mapRow = (row: AdvanceTestSlotRow): ExaminerAdvanceTestSlot => {
-  return {
-    examinerId: row.individual_id,
-    advanceTestSlot: {
-      slotDetail: {
-        slotId: row.slot_id,
-        start: row.start_time,
-        duration: row.minutes,
-      },
-      testCentre: {
-        centreId: row.tc_id,
-        centreName: row.tc_name,
-        costCode: row.tc_cost_centre_code,
-      },
-      vehicleSlotType: row.short_vst_desc,
-    },
-  };
 };
