+++ src/functions/pollJournals/framework/repo/mysql/non-test-activity-repository.ts
@@ -2,6 +2,7 @@ import * as mysql from 'mysql';
 import { query } from '../database';
 import { ExaminerNonTestActivity } from '../../../domain/examiner-non-test-activity';
 import * as moment from 'moment';
+import { mapRow } from './row-mappers/non-test-activity-row-mapper';
 
 export const getNonTestActivities = async (connectionPool: mysql.Pool)
   : Promise<ExaminerNonTestActivity[]> => {
@@ -23,36 +24,4 @@ export const getNonTestActivities = async (connectionPool: mysql.Pool)
     [windowStart, windowEnd, windowStart],
   );
   return res.map(mapRow);
-};
-
-interface NonTestActivityRow {
-  individual_id: number;
-  slot_id: number;
-  start_time: string;
-  minutes: number;
-  non_test_activity_code: string;
-  reason_desc: string;
-  tc_id: number;
-  tc_name: string;
-  tc_cost_centre_code: string;
-}
-
-const mapRow = (row: NonTestActivityRow): ExaminerNonTestActivity => {
-  return {
-    examinerId: row.individual_id,
-    nonTestActivity: {
-      slotDetail: {
-        slotId: row.slot_id,
-        start: row.start_time,
-        duration: row.minutes,
-      },
-      activityCode: row.non_test_activity_code,
-      activityDescription: row.reason_desc,
-      testCentre: {
-        centreId: row.tc_id,
-        centreName: row.tc_name,
-        costCode: row.tc_cost_centre_code,
-      },
-    },
-  };
 };
