+++ src/functions/pollJournals/framework/repo/mysql/personal-commitment-repository.ts
@@ -2,6 +2,7 @@ import * as mysql from 'mysql';
 import { query } from '../database';
 import { ExaminerPersonalCommitment } from '../../../domain/examiner-personal-commitment';
 import * as moment from 'moment';
+import { mapRow } from './row-mappers/personal-commitment-row-mapper';
 
 export const getPersonalCommitments = async (connectionPool: mysql.Pool)
   : Promise<ExaminerPersonalCommitment[]> => {
@@ -31,31 +32,6 @@ export const getPersonalCommitments = async (connectionPool: mysql.Pool)
     `,
     [windowStart, windowEnd, windowStart, windowEnd, windowStart]
     /* tslint:enable */
-  ) as PersonalCommitmentRow[];
+  );
   return res.map(mapRow);
-};
-
-interface PersonalCommitmentRow {
-  individual_id: number;
-  commitment_id: number;
-  start_date_time: string;
-  end_date_time: string;
-  non_test_activity_code: string;
-  reason_desc: string;
-}
-
-// TODO: Split out start/end date/time
-const mapRow = (row: PersonalCommitmentRow): ExaminerPersonalCommitment => {
-  return {
-    examinerId: row.individual_id,
-    personalCommitment: {
-      commitmentId: row.commitment_id,
-      startDate: row.start_date_time,
-      startTime: row.start_date_time,
-      endDate: row.end_date_time,
-      endTime: row.end_date_time,
-      activityCode: row.non_test_activity_code,
-      activityDescription: row.reason_desc,
-    },
-  };
 };
