+++ src/functions/pollJournals/framework/repo/mysql/row-mappers/__tests__/advance-test-slot-row-mapper.spec.ts
@@ -1,0 +1,35 @@
+import { mapRow } from '../advance-test-slot-row-mapper';
+
+describe('AdvanceTestSlot Row Mapper', () => {
+
+  it('should map an AdvanceTestSlot row to an ExaminerAdvanceTestSlot', () => {
+    const result = mapRow({
+      individual_id: 1,
+      slot_id: 123,
+      start_time: '2019-02-12 08:20:00',
+      minutes: 100,
+      tc_id: 3,
+      tc_name: 'test-tc',
+      tc_cost_centre_code: 'csc',
+      short_vst_desc: 'B57mins',
+    });
+    expect(result).toEqual(
+      {
+        examinerId: 1,
+        advanceTestSlot: {
+          slotDetail: {
+            slotId: 123,
+            start: '2019-02-12 08:20:00',
+            duration: 100,
+          },
+          testCentre: {
+            centreId: 3,
+            centreName: 'test-tc',
+            costCode: 'csc',
+          },
+          vehicleSlotType: 'B57mins',
+        },
+      },
+    );
+  });
+});
